ProjectCVRag/pyproject.toml
[project]
name = "chat-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "google-generativeai>=0.8.6",
    "gradio>=6.5.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.0",
    "load-dotenv>=0.1.0",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
]
